Fixed integer cast bug with timestamp for 32bit platforms
git-svn-id: http://android-market-license-verification.googlecode.com/svn/trunk@5 9bbedade-4383-ce3b-2565-04fab2f84983

+++ library/AndroidMarket/Licensing/ResponseData.php
@@ -133,11 +133,11 @@ class AndroidMarket_Licensing_ResponseData
     /**
      * Get the response timestamp
      *
-     * @return integer
+     * @return double
      */
     public function getTimestamp()
     {
-        return (int)$this->_timestamp;
+        return (double)$this->_timestamp;
     }
 
     /**
